named: address bulk insert v1.3.1 issues
This change-set is how I addressed the issues of bulk inserts not being
handled cleanly when they're not at the end of the query.

It works by first adding VALUES to the expression (not sure if this is
universally applicable to all supported SQL dialects) then adding a
submatch to the regular expression for the fields we want to repeat.
Using the submatch, we're able to repeat the fields that we're concerned
with to expand the insertion parameters.